rename component
initially the component would be a composition of Input and Value so we
could use as <Money.Input /> and <Money.Value /> but now we are going to
keep them separated as <MoneyInput /> and <MoneyValue />.
