+++ packages/ui/src/components/MoneyInput/MoneyInput.styles.ts
