+++ packages/ui/src/components/MoneyInput/MoneyInput.types.ts
@@ -2,7 +2,7 @@ import { ComponentPropsWithoutRef, ElementRef } from "react";
 import { NumericFormat, NumericFormatProps } from "react-number-format";
 import * as Common from "../Common";
 import { VariantProps } from "class-variance-authority";
-import { inputStyles } from "./Money.styles";
+import { inputStyles } from "../Input";
 
 type InputComponent = ElementRef<typeof NumericFormat>;
